+++ soridam-infra/src/main/java/sorisoop/soridam/infra/notification/ReviewCreatedEvent.java
@@ -3,7 +3,7 @@ package sorisoop.soridam.infra.notification;
 import java.time.LocalDateTime;
 
 import lombok.Builder;
-import sorisoop.soridam.domain.review.domain.Review;
+import sorisoop.soridam.domain.review_service.domain.Review;
 
 @Builder
 public record ReviewCreatedEvent(
